+++ rgx-core/src/compiler.rs
@@ -914,13 +914,15 @@ impl Compiler {
                 index,
                 name,
             },
-            RegexAst::Lookahead { expr, positive } => RegexAst::Lookahead {
+            RegexAst::Lookahead { expr, positive, non_atomic } => RegexAst::Lookahead {
                 expr: Box::new(Self::resolve_octal_backreferences(*expr, total_groups)),
                 positive,
+                non_atomic,
             },
-            RegexAst::Lookbehind { expr, positive } => RegexAst::Lookbehind {
+            RegexAst::Lookbehind { expr, positive, non_atomic } => RegexAst::Lookbehind {
                 expr: Box::new(Self::resolve_octal_backreferences(*expr, total_groups)),
                 positive,
+                non_atomic,
             },
             RegexAst::FlagGroup { flags, expr } => RegexAst::FlagGroup {
                 flags,
@@ -1018,13 +1020,15 @@ impl Compiler {
                 index,
                 name,
             },
-            RegexAst::Lookahead { expr, positive } => RegexAst::Lookahead {
+            RegexAst::Lookahead { expr, positive, non_atomic } => RegexAst::Lookahead {
                 expr: Box::new(Self::lower_flag_toggles(*expr)),
                 positive,
+                non_atomic,
             },
-            RegexAst::Lookbehind { expr, positive } => RegexAst::Lookbehind {
+            RegexAst::Lookbehind { expr, positive, non_atomic } => RegexAst::Lookbehind {
                 expr: Box::new(Self::lower_flag_toggles(*expr)),
                 positive,
+                non_atomic,
             },
             RegexAst::FlagGroup { flags, expr } => RegexAst::FlagGroup {
                 flags,
@@ -1238,13 +1242,15 @@ impl Compiler {
                 index,
                 name,
             },
-            RegexAst::Lookahead { expr, positive } => RegexAst::Lookahead {
+            RegexAst::Lookahead { expr, positive, non_atomic } => RegexAst::Lookahead {
                 expr: Box::new(Self::strip_extended_inner(*expr, in_x_mode)),
                 positive,
+                non_atomic,
             },
-            RegexAst::Lookbehind { expr, positive } => RegexAst::Lookbehind {
+            RegexAst::Lookbehind { expr, positive, non_atomic } => RegexAst::Lookbehind {
                 expr: Box::new(Self::strip_extended_inner(*expr, in_x_mode)),
                 positive,
+                non_atomic,
             },
             RegexAst::Conditional {
                 condition,
@@ -1348,13 +1354,15 @@ impl Compiler {
                 index,
                 name,
             },
-            RegexAst::Lookahead { expr, positive } => RegexAst::Lookahead {
+            RegexAst::Lookahead { expr, positive, non_atomic } => RegexAst::Lookahead {
                 expr: Box::new(Self::retarget_quantifiers_on_transparent(*expr)),
                 positive,
+                non_atomic,
             },
-            RegexAst::Lookbehind { expr, positive } => RegexAst::Lookbehind {
+            RegexAst::Lookbehind { expr, positive, non_atomic } => RegexAst::Lookbehind {
                 expr: Box::new(Self::retarget_quantifiers_on_transparent(*expr)),
                 positive,
+                non_atomic,
             },
             RegexAst::FlagGroup { flags, expr } => RegexAst::FlagGroup {
                 flags,
@@ -1451,13 +1459,15 @@ impl Compiler {
                 index,
                 name,
             }),
-            RegexAst::Lookahead { expr, positive } => Ok(RegexAst::Lookahead {
+            RegexAst::Lookahead { expr, positive, non_atomic } => Ok(RegexAst::Lookahead {
                 expr: Box::new(Self::lower_extended_char_classes(*expr)?),
                 positive,
+                non_atomic,
             }),
-            RegexAst::Lookbehind { expr, positive } => Ok(RegexAst::Lookbehind {
+            RegexAst::Lookbehind { expr, positive, non_atomic } => Ok(RegexAst::Lookbehind {
                 expr: Box::new(Self::lower_extended_char_classes(*expr)?),
                 positive,
+                non_atomic,
             }),
             RegexAst::FlagGroup { flags, expr } => Ok(RegexAst::FlagGroup {
                 flags,
@@ -2074,22 +2084,24 @@ impl Compiler {
             RegexAst::Group {
                 expr, kind, name, ..
             } => Self::assign_capture_indices_group(*expr, kind, name, next_group),
-            RegexAst::Lookahead { expr, positive } => {
+            RegexAst::Lookahead { expr, positive, non_atomic } => {
                 let (expr, next) = Self::assign_capture_indices_inner(*expr, next_group);
                 (
                     RegexAst::Lookahead {
                         expr: Box::new(expr),
                         positive,
+                non_atomic,
                     },
                     next,
                 )
             }
-            RegexAst::Lookbehind { expr, positive } => {
+            RegexAst::Lookbehind { expr, positive, non_atomic } => {
                 let (expr, next) = Self::assign_capture_indices_inner(*expr, next_group);
                 (
                     RegexAst::Lookbehind {
                         expr: Box::new(expr),
                         positive,
+                non_atomic,
                     },
                     next,
                 )
@@ -2332,24 +2344,26 @@ impl Compiler {
                 opened_groups,
                 total_groups,
             ),
-            RegexAst::Lookahead { expr, positive } => {
+            RegexAst::Lookahead { expr, positive, non_atomic } => {
                 let (expr, opened_after_expr) =
                     Self::resolve_relative_conditionals_inner(*expr, opened_groups, total_groups)?;
                 Ok((
                     RegexAst::Lookahead {
                         expr: Box::new(expr),
                         positive,
+                non_atomic,
                     },
                     opened_after_expr,
                 ))
             }
-            RegexAst::Lookbehind { expr, positive } => {
+            RegexAst::Lookbehind { expr, positive, non_atomic } => {
                 let (expr, opened_after_expr) =
                     Self::resolve_relative_conditionals_inner(*expr, opened_groups, total_groups)?;
                 Ok((
                     RegexAst::Lookbehind {
                         expr: Box::new(expr),
                         positive,
+                non_atomic,
                     },
                     opened_after_expr,
                 ))
@@ -2692,21 +2706,23 @@ impl Compiler {
                 )?),
                 quantifier,
             }),
-            RegexAst::Lookahead { expr, positive } => Ok(RegexAst::Lookahead {
+            RegexAst::Lookahead { expr, positive, non_atomic } => Ok(RegexAst::Lookahead {
                 expr: Box::new(Self::resolve_recursion_conditionals(
                     *expr,
                     total_groups,
                     named_groups,
                 )?),
                 positive,
+                non_atomic,
             }),
-            RegexAst::Lookbehind { expr, positive } => Ok(RegexAst::Lookbehind {
+            RegexAst::Lookbehind { expr, positive, non_atomic } => Ok(RegexAst::Lookbehind {
                 expr: Box::new(Self::resolve_recursion_conditionals(
                     *expr,
                     total_groups,
                     named_groups,
                 )?),
                 positive,
+                non_atomic,
             }),
             RegexAst::Group {
                 expr,